ref: move basic file operations to file-operations.js
